Feat: properly switch between settings and chat widget
