+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@emailjs/browser": "^4.4.1",
+    "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
     "@googleworkspace/meet-addons": "^1.2.0",
     "@mui/base": "5.0.0-beta.70",
