+++ styles/globals.css
@@ -22,6 +22,13 @@ body {
   padding: 0;
   overflow-x: hidden;
 }
+
+#__next,
+html,
+body {
+  height: 100%;
+}
+
 @keyframes spin {
   to {
     transform: rotate(360deg);
